python/requirements-workers.txt
# Existing dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers
torch
numpy
pydantic

# Phase 2 - Worker dependencies
apscheduler==3.10.4          # Job scheduling
anthropic==0.18.0            # Claude API
spacy==3.7.2                 # NER and NLP
scikit-learn==1.4.0          # TF-IDF, clustering
python-dotenv==1.0.0         # Environment variables
redis==5.0.1                 # Optional: Job queue (if using Redis)
psutil==5.9.7                # System monitoring
tenacity==8.2.3              # Retry logic

# Code parsing
tree-sitter==0.20.4          # Code AST parsing
pygments==2.17.2             # Syntax highlighting/parsing

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1                 # Progress bars
